+++ src/lib.rs
@@ -7,6 +7,7 @@
 //! - Enable/disable the device.
 //! - Enable/disable the RGB converter.
 //! - Set RGB converter gain.
+//! - Enable/disable the RGB converter interrupt generation.
 //! - Set the number of integration cycles.
 //! - Enable/disable the wait feature.
 //! - Set the number of wait time cycles.
@@ -131,6 +132,7 @@ impl BitFlags {
     const POWER_ON     : u8 = 0b0000_0001; // PON
     const RGBC_EN      : u8 = 0b0000_0010; // AEN
     const WAIT_EN      : u8 = 0b0000_1000; // WEN
+    const RGBC_INT_EN  : u8 = 0b0001_0000; // AIEN
     const RGBC_VALID   : u8 = 0b0000_0001; // AVALID
     const WLONG        : u8 = 0b0000_0010;
 }